pyproject.toml
[project]
name = "azure-openai-agent"
version = "0.1.0"
description = "An agentic framework for Azure OpenAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.12.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.8.0",
]
